+++ lib/src/utils/datetime_util.dart
@@ -4,17 +4,20 @@ extension DateExtractor on DateTime {
   bool isSameDate(DateTime other) =>
       year == other.year && month == other.month && day == other.day;
 
-  DateTime extractDateWith([
+  DateTime extractDateWith({
+    int? year,
+    int? month,
+    int? day,
     int hour = 0,
     int minute = 0,
     int second = 0,
     int millisecond = 0,
     int microsecond = 0,
-  ]) =>
+  }) =>
       DateTime(
-        year,
-        month,
-        day,
+        year ?? this.year,
+        month ?? this.month,
+        day ?? this.day,
         hour,
         minute,
         second,